# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Oficina Backend API",
        "description": "Documentacao base da API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/order-services": {
            "get": {
                "tags": [
                    "OrderServices"
                ],
                "summary": "Listar Ordens de Serviço",
                "operationId": "64ae0993ef3d66ab70ba8cb8b3cea167",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "created",
                                "in_analysis",
                                "pending_approval",
                                "in_renegotiation",
                                "approved",
                                "rejected",
                                "in_execution",
                                "execution_finished",
                                "delivered_and_finalized"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de OS, ordenada por prioridade de status (Execução > Aguardando Aprovação > Diagnóstico > Recebida) e mais antigas primeiro. Sem filtro de status, OS rejeitadas/finalizadas/entregues ficam ocultas (exclusão lógica)."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "OrderServices"
                ],
                "summary": "Criar nova OS (Atendente). Status inicial: CREATED",
                "operationId": "576a042dd94d7633caf14541c19c887b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "customer_id",
                                    "vehicle_id",
                                    "complaint"
                                ],
                                "properties": {
                                    "customer_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "vehicle_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "complaint": {
                                        "type": "string",
                                        "example": "Carro fazendo barulho ao frear"
                                    },
                                    "mechanic_user_id": {
                                        "type": "integer",
                                        "example": null,
                                        "nullable": true
                                    },
                                    "services": {
                                        "description": "Opcional. Serviços solicitados pelo cliente na abertura — sem preço, apenas informativo. O orçamento oficial é gerado depois via generate-budget.",
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "service_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "parts": {
                                        "description": "Opcional. Peças solicitadas pelo cliente na abertura — sem preço, apenas informativo. O orçamento oficial é gerado depois via generate-budget.",
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "part_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OS criada (status: created). Se services/parts forem enviados, aparecem em requested_services/requested_parts na resposta, sem preço e sem gerar orçamento."
                    },
                    "404": {
                        "description": "Cliente, veículo, serviço ou peça solicitada não encontrado"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/order-services/{id}": {
            "get": {
                "tags": [
                    "OrderServices"
                ],
                "summary": "Buscar OS por ID",
                "operationId": "8f47ee4d3c39655506d65e17dc96ee41",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados da OS, incluindo public_status (status simplificado de 6 estados) e requested_services/requested_parts (itens solicitados na abertura, sem preço)."
                    },
                    "404": {
                        "description": "Não encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/order-services/{id}/send-to-analysis": {
            "post": {
                "tags": [
                    "OrderServices"
                ],
                "summary": "Enviar OS para análise (Mecânico). CREATED → IN_ANALYSIS",
                "operationId": "96e7a70f3fa9c069438479f6f0949e95",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status atualizado para IN_ANALYSIS"
                    },
                    "404": {
                        "description": "Não encontrada"
                    },
                    "422": {
                        "description": "Transição inválida"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/order-services/{id}/generate-budget": {
            "post": {
                "tags": [
                    "OrderServices"
                ],
                "summary": "Gerar orçamento (Mecânico). IN_ANALYSIS → PENDING_APPROVAL",
                "operationId": "a6f63e607820fdbcd7895b64bc069e95",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "services",
                                    "parts"
                                ],
                                "properties": {
                                    "services": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "service_id": {
                                                    "type": "integer"
                                                },
                                                "quantity": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "parts": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "part_id": {
                                                    "type": "integer"
                                                },
                                                "quantity": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "notes": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Orçamento gerado, status PENDING_APPROVAL"
                    },
                    "404": {
                        "description": "Não encontrada"
                    },
                    "422": {
                        "description": "Transição inválida ou dados inválidos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/order-services/{id}/approve-budget": {
            "post": {
                "tags": [
                    "OrderServices"
                ],
                "summary": "Cliente aprova orçamento (Atendente). PENDING_APPROVAL → APPROVED",
                "operationId": "9250bc51d0fc7770cd01c2c744f93779",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status atualizado para APPROVED"
                    },
                    "404": {
                        "description": "Não encontrada"
                    },
                    "422": {
                        "description": "Transição inválida"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token gerado pelo endpoint POST /api/auth/login",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}